renamed class, file and spec for api lookup data to dropdown_data (from seed)
