+++ app/interfaces/api/v1/advocates/root.rb
@@ -11,7 +11,7 @@ module API
         mount API::V1::Advocates::Expense
         mount API::V1::Advocates::DateAttended
         mount API::V1::Advocates::RepresentationOrder
-        mount API::V1::Advocates::Seed
+        mount API::V1::DropdownData
         add_swagger_documentation(
           api_version: "v1",
           hide_documentation_path: true,
