+++ spec/factories/claims.rb
@@ -49,7 +49,7 @@ FactoryGirl.define do
     case_number { Faker::Number.number(10) }
     advocate
     source { 'web' }
-    apply_vat  true
+    apply_vat  false
     after(:build) do |claim|
       claim.creator = claim.advocate
     end
